# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: API для управления банковскими картами, пользователями и аутентификации
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8091
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /users:
    post:
      tags:
        - User Management
      summary: Создать пользователя
      description: Создает нового пользователя в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Неверные данные запроса
          content:
            text/plain:
              schema:
                type: string
                example: "Пользователь с таким email уже существует"

    get:
      tags:
        - User Management
      summary: Получить всех пользователей
      description: Возвращает список всех пользователей системы
      responses:
        '200':
          description: Список пользователей успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
  /users/{id}:
    get:
      tags:
        - User Management
      summary: Получить пользователя по ID
      description: Возвращает информацию о пользователе по его идентификатору
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Пользователь не найден

    put:
      tags:
        - User Management
      summary: Обновить пользователя
      description: Обновляет информацию о пользователе по его идентификатору
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Пользователь успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Неверные данные запроса
          content:
            text/plain:
              schema:
                type: string
                example: "Email уже используется другим пользователем"
        '404':
          description: Пользователь не найден

    delete:
      tags:
        - User Management
      summary: Удалить пользователя
      description: Удаляет пользователя по его идентификатору
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Пользователь успешно удален
          content:
            text/plain:
              schema:
                type: string
                example: "Пользователь удален"
        '400':
          description: Ошибка при удалении пользователя
          content:
            text/plain:
              schema:
                type: string
                example: "Невозможно удалить пользователя с привязанными картами"
        '404':
          description: Пользователь не найден
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDto:
      type: object
      required:
        - id
        - email
        - phoneNumber
        - firstName
        - lastName
        - role
        - birthDate
        - fullName
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор пользователя
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "user@example.com"
        phoneNumber:
          type: string
          description: Номер телефона
          example: "+79991234567"
        firstName:
          type: string
          description: Имя пользователя
          example: "Иван"
          maxLength: 50
        lastName:
          type: string
          description: Фамилия пользователя
          example: "Иванов"
          maxLength: 50
        role:
          $ref: '#/components/schemas/Role'
        birthDate:
          type: string
          format: date
          description: Дата рождения
          example: "1990-01-15"
        fullName:
          type: string
          description: Полное имя пользователя
          example: "Иван Иванов"
    CreateUserRequest:
      type: object
      required:
        - email
        - phoneNumber
        - password
        - firstName
        - lastName
        - role
        - birthDate
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: "user@example.com"
          maxLength: 100
        phoneNumber:
          type: string
          description: Номер телефона в международном формате
          example: "+79991234567"
        password:
          type: string
          format: password
          description: Пароль пользователя
          example: "Secure123"
        firstName:
          type: string
          description: Имя пользователя
          example: "Иван"
          maxLength: 50
        lastName:
          type: string
          description: Фамилия пользователя
          example: "Иванов"
          maxLength: 50
        role:
          $ref: '#/components/schemas/Role'
        birthDate:
          type: string
          format: date
          description: Дата рождения
          example: "1990-01-15"
    UpdateUserRequest:
      type: object
      description: Запрос на обновление данных пользователя. Все поля опциональны.
      properties:
        phoneNumber:
          type: string
          description: Новый номер телефона в международном формате
          example: "+79991234567"
          nullable: true
        firstName:
          type: string
          description: Новое имя пользователя
          example: "Иван"
          maxLength: 50
          nullable: true
        lastName:
          type: string
          description: Новая фамилия пользователя
          example: "Иванов"
          maxLength: 50
          nullable: true
        birthDate:
          type: string
          format: date
          description: Новая дата рождения пользователя
          example: "1990-01-15"
          nullable: true
...
    Role:
      type: string
      enum: [ADMIN, USER]
      description: Статус банковской карты